fixes #5, #6, #7
fix #5 - altcoins parsing problem

The logic for adding new altcoins to the list of altcoins was using the
length of the first value as the index into the array instead of the
length of the array itself.

fix #6 - ticker delay should be configurable

Added a -T option for the ticker time to the balances script and then
passed the value to the invocation of tickline, which already supported
the option, it just needed to be parsed and passed in.

fix #7 - Altcoins should allow display in other currency, too

The "${CURRENCY}" value was not being passed to the invocation of
`get_prices`.  This is now fixed, the result is that you can actually
leverage this to do things like show the ratio of STEEM to SBD.  For
example, `balances.sh -a STEEM -c SBD` and put it in a ticker, and you
can now see when the best time is to convert that spare SBD into STEEM.

Other changes:

 * Made it easier to tell where the altcoins were in the output by
placing brackets around them.  When displaying worth and balances it was
hard to tell when you were seeing the value of account information
versus just displaying an SBD or STEEM coin value via `-a`.